+++ superwall/src/main/java/com/superwall/sdk/analytics/internal/trackable/TrackableSuperwallEvent.kt
@@ -596,7 +596,9 @@ sealed class InternalSuperwallEvent(
         sealed class State {
             class Start : State()
 
-            class Fail : State()
+            data class Fail(
+                val errorMessage: String,
+            ) : State()
 
             class Timeout : State()
 
@@ -619,6 +621,7 @@ sealed class InternalSuperwallEvent(
                     is PaywallWebviewLoad.State.Fail ->
                         SuperwallEvent.PaywallWebviewLoadFail(
                             paywallInfo,
+                            state.errorMessage,
                         )
 
                     is PaywallWebviewLoad.State.Complete ->